requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
docker-pycreds==0.4.0
easydict==1.10
filelock==3.6.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.1
huggingface-hub>=0.6.0
idna==3.3
imageio==2.19.1
importlib-metadata==4.11.3
joblib==1.1.0
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
multidict==6.0.2
networkx==2.8
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
Pillow==9.1.0
promise==2.3
protobuf==3.20.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.3
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.53
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
tokenizers>=0.10.3
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchmetrics==0.8.2
torchvision==0.12.0+cu113
tqdm==4.64.0
transformers==4.24.0
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.12.16
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
datasets==2.6.1
flamingo-pytorch==0.1.2
sentencepiece==0.1.97
